Improve mutation observer handling
A change was made to the mutation observer's behavior to prevent an infinite loop caused when a 'hidden' input type with a model directive has its value altered via Javascript. Previously, such a change would update the 'value' attribute which in turn would trigger the mutation observer and re-run the model effect, resulting in a never-ending loop. This update ensures the directive only runs its effect when necessary, thereby preventing the infinite loop from occurring.